requirements.txt
# Core FastAPI stack
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database (replaces ActiveRecord)
sqlalchemy==2.0.23
alembic==1.13.1
mysql-connector-python==8.2.0
pymysql==1.1.0

# Authentication & Security (replaces Devise/CanCanCan)
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Background Jobs (replaces Sidekiq)
celery==5.3.4
redis==5.0.1

# Search (replaces Rails Elasticsearch integration)
elasticsearch==8.11.1

# HTTP Client (replaces RestClient)
httpx==0.25.2
requests==2.31.0

# Email (replaces ActionMailer)
fastapi-mail==1.4.1
jinja2==3.1.2

# File Upload & Storage (replaces ActiveStorage)
python-multipart==0.0.6
boto3==1.34.0  # For S3 storage

# Validation & Serialization (replaces Rails validations)
pydantic[email]==2.5.0
email-validator==2.1.0

# Rate Limiting (replaces Rack::Attack)
slowapi==0.1.9

# Pagination (replaces will_paginate)
fastapi-pagination==0.12.14

# Background Scheduling (replaces whenever gem)
# celery-beat is part of celery, no separate package needed
python-crontab==3.0.0

# Monitoring & Logging (replaces Rails logging)
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Configuration Management
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
factory-boy==3.3.0  # replaces factory_bot
faker==20.1.0

# Database Fixtures & Seeds (replaces Rails seeds)
factory-boy==3.3.0

# JSON Processing (enhanced)
orjson==3.9.10  # Faster JSON

# Date/Time handling (replaces Rails Time zones)
python-dateutil==2.8.2
pytz==2023.3

# Encryption (replaces attr_encrypted)
cryptography==41.0.7

# PDF Generation (if needed)
reportlab==4.0.7

# Excel/CSV processing (replaces Rails CSV)
pandas==2.1.4
openpyxl==3.1.2

# Image processing (if needed)
pillow==10.1.0

# Async database support
asyncpg==0.29.0  # For PostgreSQL if needed later
aiomysql==0.2.0  # For async MySQL

# API Documentation
Markdown==3.5.1